cherry pick from master

+++ src/string.c
@@ -1,9 +1,7 @@
 #include "types.h"
 #include "utils.h"
 
-void*
-memset(void *dst, int c, uint n)
-{
+void *memset(void *dst, int c, uint n) {
   char *cdst = (char *) dst;
   int i;
   for(i = 0; i < n; i++){
@@ -12,9 +10,7 @@ memset(void *dst, int c, uint n)
   return dst;
 }
 
-int
-memcmp(const void *v1, const void *v2, uint n)
-{
+int memcmp(const void *v1, const void *v2, uint n) {
   const uchar *s1, *s2;
 
   s1 = v1;
@@ -28,9 +24,7 @@ memcmp(const void *v1, const void *v2, uint n)
   return 0;
 }
 
-void*
-memmove(void *dst, const void *src, size_t n)
-{
+void *memmove(void *dst, const void *src, uint n) {
   const char *s;
   char *d;
 
@@ -52,15 +46,11 @@ memmove(void *dst, const void *src, size_t n)
 }
 
 // memcpy exists to placate GCC.  Use memmove.
-void*
-memcpy(void *dst, const void *src, uint n)
-{
+void *memcpy(void *dst, const void *src, uint n) {
   return memmove(dst, src, n);
 }
 
-int
-strncmp(const char *p, const char *q, uint n)
-{
+int strncmp(const char *p, const char *q, uint n) {
   while(n > 0 && *p && *p == *q)
     n--, p++, q++;
   if(n == 0)
@@ -68,9 +58,7 @@ strncmp(const char *p, const char *q, uint n)
   return (uchar)*p - (uchar)*q;
 }
 
-char*
-strncpy(char *s, const char *t, int n)
-{
+char *strncpy(char *s, const char *t, int n) {
   char *os;
 
   os = s;
@@ -82,9 +70,7 @@ strncpy(char *s, const char *t, int n)
 }
 
 // Like strncpy but guaranteed to NUL-terminate.
-char*
-safestrcpy(char *s, const char *t, int n)
-{
+char *safestrcpy(char *s, const char *t, int n) {
   char *os;
 
   os = s;
@@ -96,9 +82,7 @@ safestrcpy(char *s, const char *t, int n)
   return os;
 }
 
-int
-strlen(const char *s)
-{
+int strlen(const char *s) {
   int n;
 
   for(n = 0; s[n]; n++)
